fix: resolve img path error and modal attribute
