+++ frontend/package.json
@@ -28,6 +28,7 @@
     "@iconify-json/fa6-regular": "^1.1.6",
     "@iconify-json/fa6-solid": "^1.1.7",
     "@iconify-json/fluent": "^1.1.16",
+    "@iconify-json/fxemoji": "^1.1.2",
     "@iconify-json/iconoir": "^1.1.11",
     "@iconify-json/ri": "^1.1.3",
     "@tailwindcss/forms": "^0.5.3",
